Create Etch and Flood
This simple script is used to first take in a set of masses then iterates in time to etch away at different cells based on the material content. These rates can be tunable and the structure allows for multiple materials to be used in this simulation. After a material cell is completely etched away, it is flooded with etchant.